+++ src/states/states.cpp
@@ -1,0 +1,5 @@
+#include "states.h"
+
+// this will hold all the data relevent to the status of the car
+// blinkerState blinkers;
+driveState drive;
